requirements.txt
opencv-python>=4.5.5.64
numpy>=1.22.3
PyAutoGUI>=0.9.53
Pillow>=9.1.0
pytest>=7.0.1
pytest-cov>=3.0.0
pyyaml>=6.0
pytesseract>=0.3.9
# tkinter is included with Python standard library - no need to install separately
